drop null values from CloudFormation responses
for example, AWS specifies that the "Reason" field is optional for
successful responses, but if it's present, it must be a string. since
null is not a string, the entire response is discarded as a failure if
it's present.

this commit also adds property testing for the CloudFormation custom
resource responses in general.