add PhysicalResourceId to CloudFormation Custom Resource interface, where appropriate
PhysicalResourceId is "always sent with Update and Delete requests;
never sent with Create" (according to AWS docs) and is commonly
required to actually perform the work needed to do an update or
delete of the resource. This PR encodes that contract and makes life a
bit easier for developers implementing update/delete, since they won't
have to deal with the optionality of the value in the event encoding.